StreamLit_App/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
requests
streamlit
